lxd/storage: Remove legacy btrfs implementation
