+++ lxd/patches_utils.go
@@ -1,13 +1,23 @@
 package main
 
 import (
+	"fmt"
 	"os"
+	"path"
+	"path/filepath"
+	"sort"
+	"strings"
+
+	"golang.org/x/sys/unix"
 
 	"github.com/lxc/lxd/lxd/project"
+	"github.com/lxc/lxd/lxd/state"
 	driver "github.com/lxc/lxd/lxd/storage"
 	"github.com/lxc/lxd/shared"
+	"github.com/lxc/lxd/shared/logger"
 )
 
+// For 'dir' storage backend.
 func dirSnapshotDeleteInternal(projectName, poolName string, snapshotName string) error {
 	snapshotContainerMntPoint := driver.GetSnapshotMountPoint(projectName, poolName, snapshotName)
 	if shared.PathExists(snapshotContainerMntPoint) {
@@ -36,4 +46,239 @@ func dirSnapshotDeleteInternal(projectName, poolName string, snapshotName string
 	}
 
 	return nil
+}
+
+// For 'btrfs' storage backend.
+func btrfsSubVolumeCreate(subvol string) error {
+	parentDestPath := filepath.Dir(subvol)
+	if !shared.PathExists(parentDestPath) {
+		err := os.MkdirAll(parentDestPath, 0711)
+		if err != nil {
+			return err
+		}
+	}
+
+	_, err := shared.RunCommand(
+		"btrfs",
+		"subvolume",
+		"create",
+		subvol)
+	if err != nil {
+		logger.Errorf("Failed to create BTRFS subvolume \"%s\": %v", subvol, err)
+		return err
+	}
+
+	return nil
+}
+
+func btrfsSnapshotDeleteInternal(projectName, poolName string, snapshotName string) error {
+	snapshotSubvolumeName := driver.GetSnapshotMountPoint(projectName, poolName, snapshotName)
+	// Also delete any leftover .ro snapshot.
+	roSnapshotSubvolumeName := fmt.Sprintf("%s.ro", snapshotSubvolumeName)
+	names := []string{snapshotSubvolumeName, roSnapshotSubvolumeName}
+	for _, name := range names {
+		if shared.PathExists(name) && btrfsIsSubVolume(name) {
+			err := btrfsSubVolumesDelete(name)
+			if err != nil {
+				return err
+			}
+		}
+	}
+
+	sourceSnapshotMntPoint := shared.VarPath("snapshots", project.Prefix(projectName, snapshotName))
+	os.Remove(sourceSnapshotMntPoint)
+	os.Remove(snapshotSubvolumeName)
+
+	sourceName, _, _ := shared.InstanceGetParentAndSnapshotName(snapshotName)
+	snapshotSubvolumePath := driver.GetSnapshotMountPoint(projectName, poolName, sourceName)
+	os.Remove(snapshotSubvolumePath)
+	if !shared.PathExists(snapshotSubvolumePath) {
+		snapshotMntPointSymlink := shared.VarPath("snapshots", project.Prefix(projectName, sourceName))
+		os.Remove(snapshotMntPointSymlink)
+	}
+
+	return nil
+}
+
+func btrfsSubVolumeQGroup(subvol string) (string, error) {
+	output, err := shared.RunCommand(
+		"btrfs",
+		"qgroup",
+		"show",
+		"-e",
+		"-f",
+		subvol)
+
+	if err != nil {
+		return "", fmt.Errorf("Quotas disabled on filesystem")
+	}
+
+	var qgroup string
+	for _, line := range strings.Split(output, "\n") {
+		if line == "" || strings.HasPrefix(line, "qgroupid") || strings.HasPrefix(line, "---") {
+			continue
+		}
+
+		fields := strings.Fields(line)
+		if len(fields) != 4 {
+			continue
+		}
+
+		qgroup = fields[0]
+	}
+
+	if qgroup == "" {
+		return "", fmt.Errorf("Unable to find quota group")
+	}
+
+	return qgroup, nil
+}
+
+func btrfsSubVolumeDelete(subvol string) error {
+	// Attempt (but don't fail on) to delete any qgroup on the subvolume
+	qgroup, err := btrfsSubVolumeQGroup(subvol)
+	if err == nil {
+		shared.RunCommand(
+			"btrfs",
+			"qgroup",
+			"destroy",
+			qgroup,
+			subvol)
+	}
+
+	// Attempt to make the subvolume writable
+	shared.RunCommand("btrfs", "property", "set", subvol, "ro", "false")
+
+	// Delete the subvolume itself
+	_, err = shared.RunCommand(
+		"btrfs",
+		"subvolume",
+		"delete",
+		subvol)
+
+	return err
+}
+
+func btrfsSubVolumesDelete(subvol string) error {
+	// Delete subsubvols.
+	subsubvols, err := btrfsSubVolumesGet(subvol)
+	if err != nil {
+		return err
+	}
+	sort.Sort(sort.Reverse(sort.StringSlice(subsubvols)))
+
+	for _, subsubvol := range subsubvols {
+		err := btrfsSubVolumeDelete(path.Join(subvol, subsubvol))
+		if err != nil {
+			return err
+		}
+	}
+
+	// Delete the subvol itself
+	err = btrfsSubVolumeDelete(subvol)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func btrfsSnapshot(s *state.State, source string, dest string, readonly bool) error {
+	var output string
+	var err error
+	if readonly && !s.OS.RunningInUserNS {
+		output, err = shared.RunCommand(
+			"btrfs",
+			"subvolume",
+			"snapshot",
+			"-r",
+			source,
+			dest)
+	} else {
+		output, err = shared.RunCommand(
+			"btrfs",
+			"subvolume",
+			"snapshot",
+			source,
+			dest)
+	}
+	if err != nil {
+		return fmt.Errorf(
+			"subvolume snapshot failed, source=%s, dest=%s, output=%s",
+			source,
+			dest,
+			output,
+		)
+	}
+
+	return err
+}
+
+func btrfsIsSubVolume(subvolPath string) bool {
+	fs := unix.Stat_t{}
+	err := unix.Lstat(subvolPath, &fs)
+	if err != nil {
+		return false
+	}
+
+	// Check if BTRFS_FIRST_FREE_OBJECTID
+	if fs.Ino != 256 {
+		return false
+	}
+
+	return true
+}
+
+func btrfsSubVolumeIsRo(path string) bool {
+	output, err := shared.RunCommand("btrfs", "property", "get", "-ts", path)
+	if err != nil {
+		return false
+	}
+
+	return strings.HasPrefix(string(output), "ro=true")
+}
+
+func btrfsSubVolumeMakeRo(path string) error {
+	_, err := shared.RunCommand("btrfs", "property", "set", "-ts", path, "ro", "true")
+	return err
+}
+
+func btrfsSubVolumeMakeRw(path string) error {
+	_, err := shared.RunCommand("btrfs", "property", "set", "-ts", path, "ro", "false")
+	return err
+}
+
+func btrfsSubVolumesGet(path string) ([]string, error) {
+	result := []string{}
+
+	if !strings.HasSuffix(path, "/") {
+		path = path + "/"
+	}
+
+	// Unprivileged users can't get to fs internals
+	filepath.Walk(path, func(fpath string, fi os.FileInfo, err error) error {
+		// Skip walk errors
+		if err != nil {
+			return nil
+		}
+
+		// Ignore the base path
+		if strings.TrimRight(fpath, "/") == strings.TrimRight(path, "/") {
+			return nil
+		}
+
+		// Subvolumes can only be directories
+		if !fi.IsDir() {
+			return nil
+		}
+
+		// Check if a btrfs subvolume
+		if btrfsIsSubVolume(fpath) {
+			result = append(result, strings.TrimPrefix(fpath, path))
+		}
+
+		return nil
+	})
+
+	return result, nil
 }
